# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the server"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/tools": {
      "get": {
        "summary": "Fetch all tools",
        "tags": [
          "Tools"
        ],
        "description": "Retrieve all tools with optional pagination.",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "The page number for pagination."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "The number of items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of tools with pagination details."
          },
          "400": {
            "description": "No tools found."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "post": {
        "summary": "Create a new tool",
        "tags": [
          "Tools"
        ],
        "description": "Add a new tool to the database.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "descriptionUnique": {
                    "type": "string"
                  },
                  "logoOwners": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "isOpenSource": {
                    "type": "boolean"
                  },
                  "sourceCodeUrl": {
                    "type": "string"
                  },
                  "linkOurGuide": {
                    "type": "string"
                  },
                  "linkOurVideo": {
                    "type": "string"
                  },
                  "topic": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tool created successfully."
          },
          "400": {
            "description": "Missing required fields or validation error."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/tools/{id}": {
      "put": {
        "summary": "Update an existing tool",
        "tags": [
          "Tools"
        ],
        "description": "Modify the details of an existing tool.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the tool to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "descriptionUnique": {
                    "type": "string"
                  },
                  "logoOwners": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "isOpenSource": {
                    "type": "boolean"
                  },
                  "sourceCodeUrl": {
                    "type": "string"
                  },
                  "linkOurGuide": {
                    "type": "string"
                  },
                  "linkOurVideo": {
                    "type": "string"
                  },
                  "topic": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool updated successfully."
          },
          "400": {
            "description": "Missing required fields or validation error."
          },
          "404": {
            "description": "Tool not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "summary": "Delete an existing tool",
        "tags": [
          "Tools"
        ],
        "description": "Remove a tool from the database.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the tool to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Tool deleted successfully."
          },
          "404": {
            "description": "Tool not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/topics": {
      "get": {
        "summary": "Retrieve all topics",
        "tags": [
          "Topics"
        ],
        "responses": {
          "200": {
            "description": "List of topics retrieved successfully"
          },
          "400": {
            "description": "No topics found"
          }
        }
      },
      "post": {
        "summary": "Create a new topic",
        "tags": [
          "Topics"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "iconTools": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Topic created successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "409": {
            "description": "Duplicate category or default topic already exists"
          }
        }
      }
    },
    "/api/topics/tools": {
      "get": {
        "summary": "Retrieve tools by topic",
        "tags": [
          "Topics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Topic ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Tools retrieved successfully"
          },
          "404": {
            "description": "Topic or tools not found"
          }
        }
      }
    },
    "/api/topics/{id}": {
      "put": {
        "summary": "Update an existing topic",
        "tags": [
          "Topics"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Topic ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "iconTools": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Topic updated successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Topic not found"
          },
          "409": {
            "description": "Duplicate category"
          }
        }
      },
      "delete": {
        "summary": "Delete a topic",
        "tags": [
          "Topics"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Topic ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Topic deleted successfully"
          },
          "404": {
            "description": "Topic not found"
          }
        }
      }
    },
...
  },
...
}